friendship endpoints

+++ src/routes/users.router.js
@@ -3,20 +3,38 @@ import { userService } from '../services/user.service';
 import { validatorHandler } from '../middlewares/validator.handler';
 import { dataFilterToC, dataFilterToU } from '../schemas/user.schema';
 import { login } from '../auth/authMiddlewares';
+import passport from '../auth';
+import { ac, checkAccess } from '../auth/accessControl';
+import { user } from 'pg/lib/defaults';
 const users = Router();
-
-users.get('/', async (_req, res, next) => {
-  try {
-    const users = await userService.find();
-    res.json(users);
-  } catch (err) {
-    next(err);
+//ADMIN RESOURCES
+users.get(
+  '/',
+  checkAccess((roles) => ac.can(roles).readAny('user')),
+  async ({ query: { offset, limit } }, res, next) => {
+    try {
+      const users = await userService.find({ offset, limit });
+      res.json(users);
+    } catch (err) {
+      next(err);
+    }
   }
-});
-
+);
+users.get(
+  '/:id',
+  checkAccess((roles) => ac.can(roles).readAny('user')),
+  async ({ params: { id } }, res, next) => {
+    try {
+      const user = await userService.findOne(id);
+      res.json(user);
+    } catch (err) {
+      next(err);
+    }
+  }
+);
 users.post(
   '/',
-  validatorHandler(dataFilterToU),
+  checkAccess((roles) => ac.can(roles).createAny('user')),
   async ({ body }, res, next) => {
     try {
       const user = await userService.create(body);
@@ -28,6 +46,7 @@ users.post(
 );
 users.patch(
   '/:id',
+  checkAccess((roles) => ac.can(roles).updateAny('user')),
   validatorHandler(dataFilterToU),
   async ({ params: { id }, body }, res, next) => {
     try {
@@ -38,14 +57,17 @@ users.patch(
     }
   }
 );
-users.delete('/:id', async ({ params: { id } }, res, next) => {
-  try {
-    await userService.delete(id);
-    res.json({ id });
-  } catch (err) {
-    next(err);
+users.delete(
+  '/:id',
+  checkAccess((roles) => ac.can(roles).deleteAny('user')),
+  async ({ params: { id } }, res, next) => {
+    try {
+      await userService.delete(id);
+      res.json({ id });
+    } catch (err) {
+      next(err);
+    }
   }
-});
+);
 
-users.post('/login', login);
 export default users;
